# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookverse API",
        "description": "Bookverse sosyal okuma uygulaması REST API dokümantasyonu. Token almak için /api/login veya /api/register kullanın, ardından Authorize ile Bearer token ekleyin.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Akış (görünür paylaşımlar)",
                "operationId": "cce7863441436fd49dc474491f658292",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paylaşım listesi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Yeni paylaşım oluştur",
                "operationId": "21a6ac5f9280a3cf958c3d066b5600ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "type"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Bu kitabı çok beğendim."
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "thought",
                                        "enum": [
                                            "thought",
                                            "quote"
                                        ]
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oluşturuldu"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Paylaşım detayı",
                "operationId": "282e197939666875838f96d76bb8549e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paylaşım"
                    },
                    "403": {
                        "description": "Görüntüleme yetkisi yok"
                    },
                    "404": {
                        "description": "Bulunamadı"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Paylaşımı sil",
                "operationId": "9f90bc40d15de327d974a5e0cbd3c649",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Silindi"
                    },
                    "403": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}/like": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Beğeni aç/kapa",
                "operationId": "82c2b59b3fdf6204b7085c55600421a5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beğeni durumu"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum personal access token",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}